normal updating

+++ Source/WNA/DMExtension/MassCapacity.cs
@@ -4,6 +4,6 @@ namespace WNA.DMExtension
 {
     public class MassCapacity : DefModExtension
     {
-        public int massCapacity;
+        public float massCapacity;
     }
 }
